% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/anonymize.R
\name{anonymize}
\alias{anonymize}
\title{anonymize}
\usage{
anonymize(x)
}
\arguments{
\item{x}{a list of factors}
}
\value{
a list of numbers
}
\description{
recodes a list of factors to a number from 100 to 999
}
\author{
Rollie Parrish
}

