% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/beeswarm_plot.R
\name{beeswarm_plot}
\alias{beeswarm_plot}
\title{beeswarm_plot}
\usage{
beeswarm_plot(data = NULL, title = "GAMUT Metric title")
}
\arguments{
\item{data}{dataframe from pd() function that contains program_name, ID, redcap_data_access_group, num, den, label, and mark}

\item{title}{plot title}
}
\value{
returns a beeswarm plot
}
\description{
build a horizontal beeswarm plot with error bars for individual program
}
\author{
Rollie Parrish
}

