% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{get_epa_facilities}
\alias{get_epa_facilities}
\title{get_epa_facilities}
\usage{
get_epa_facilities()
}
\description{
get_epa_facilities
}
\details{
This function imports EPA facilities data which has data on facility location and type. The function takes SIC codes and reclassifies them to new categories.
}
\author{
Sean Turner (sean.turner@pnnl.gov)
}
