% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{get_irrigation_bcm}
\alias{get_irrigation_bcm}
\title{get_irrigation_bcm}
\usage{
get_irrigation_bcm()
}
\description{
get_irrigation_bcm
}
\details{
This function loads in HUC2 irrigation data in billion cubic meters.
}
\author{
Kristian Nelson (kristian.nelson@pnnl.gov)
}
