% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{get_population}
\alias{get_population}
\title{get_population}
\usage{
get_population()
}
\description{
get_population
}
\details{
This function loads in US Census population data for the cities in gamut.
}
\author{
Sean Turner (sean.turner@pnnl.gov)
}
