% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{get_source_contribution}
\alias{get_source_contribution}
\title{get_source_contribution}
\usage{
get_source_contribution(data_dir, file_paths)
}
\arguments{
\item{data_dir}{your data directory}

\item{file_paths}{the file paths to the gamut inputs}
}
\description{
get_source_contribution
}
\details{
This function gets a city's water source contribution breakdown. This data is from the UWSCatCH dataset.
}
\author{
Sean Turner (sean.turner@pnnl.gov)
}
