% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{get_watershed_ts}
\alias{get_watershed_ts}
\title{get_watershed_ts}
\usage{
get_watershed_ts(watersheds)
}
\arguments{
\item{watersheds}{select watershed}
}
\description{
get_watershed_ts
}
\details{
This function loads in runoff time series and selects the watershed being analyzed and its time series. This information is used in the yield analysis.
}
\author{
Kristian Nelson (kristian.nelson@pnnl.gov)
}
