% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{get_watershed_usage}
\alias{get_watershed_usage}
\title{get_watershed_usage}
\usage{
get_watershed_usage(city)
}
\arguments{
\item{city}{city that is being analyzed}
}
\description{
get_watershed_usage
}
\details{
This function loads in the watershed usage table which provides information on how much a city uses a certain watershed.
}
\author{
Kristian Nelson (kristian.nelson@pnnl.gov)
}
