% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pd}
\alias{pd}
\title{pd}
\usage{
pd(...)
}
\arguments{
\item{...}{comma separated list of variables to include. First variable should be
group, which is either "program_name" or "ID".}
}
\description{
Returns a data frame with aggregated numerators and denominators.
Filters out incomplete cases and those with more denominator cases
than numerators.
}
\author{
Rollie Parrish
}

