% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_to_mysql.R
\name{send_to_mysql}
\alias{send_to_mysql}
\title{send_redcap_to_mysql.R}
\usage{
send_to_mysql()
}
\value{
a list - table has the raw data and plot has the dotplot
}
\description{
updates the GAMUT dashboard database tables
}
\author{
Rollie Parrish
}

