% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sendit.R
\name{sendit}
\alias{sendit}
\title{sendit.R}
\usage{
sendit(recipients, filename, subject, message, expireDays = 15,
  emailFrom = 1, url = "http://httpbin.org/post", username, password)
}
\arguments{
\item{recipients}{a comma-separated list of email addresses}

\item{filename}{the file to be sent}

\item{subject}{email subject}

\item{message}{email message}

\item{expireDays}{number of days that the file will be available. Default is 7.}

\item{emailFrom}{index number the sending email address for
the specified REDCapR username. Default is 1 (first one)}

\item{url}{address of the SenditAPI module}

\item{username}{}

\item{password}{}
}
\description{
Wrapper function for httr that sends PDF report to a list of recipients.
Requires REDCap Sendit API module and a valid REDCap user account.
}
\author{
Rollie Parrish
}

