% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAMUT_data.R
\name{GAMUT_data}
\alias{GAMUT_data}
\title{GAMUT_Data}
\usage{
GAMUT_data(file = "data/GAMUT.Rdata")
}
\description{
get GAMUT data from REDCap, transform/aggregate by
individual program. Save as metricData to
GAMUT.Rdata file in the data folder
}
\author{
Rollie Parrish
}

