% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd.R
\name{pd}
\alias{pd}
\title{pd}
\usage{
pd(dag = "", operator = "", ...)
}
\arguments{
\item{dag}{list of redcap data access groups. Shows the program name for the
data access groups in the list and the anonymous ID for all others. If empty (default), then all
program names will be shown.}

\item{operator}{used to show multiple data access groups from the same operator (ie. AIM, MTR, etc.).
Will display the program names when the first three letters of the program name matches the value
in this field.}

\item{...}{comma separated list of variables to include.}
}
\description{
Returns a data frame with aggregated numerators and denominators.
Filters out incomplete cases and those with more denominator cases
than numerators.
}
\author{
Rollie Parrish
}

