% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_metrics.R
\name{plot_metrics}
\alias{plot_metrics}
\title{plot_metrics}
\usage{
plot_metrics(metric = "metric name", program_name, num, den,
  reordering = "decreasing", qlabel = "Estimated rate", rate = "\%", ...)
}
\arguments{
\item{metric}{Name of the GAMUT metric}

\item{program_name}{Vector of names of the transport program}

\item{num}{vector of numerators for the metric}

\item{den}{vector of denominators}

\item{reordering}{Sets the sorting order. Default is 'decreasing'}

\item{qlabel}{Default is 'Estimated rate'}

\item{rate}{units of measure. Default is '\%'}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Returns a dot plot for a given metric. Based on numerators and denominators
}
\author{
Rollie Parrish
}

